refactor: рефакторинг страницы Stories по замечаниям
